Don't use threads on threadsForRuns @ 1. Bug fixes

+++ proj2/src/headers/vctQuant.h
@@ -30,7 +30,7 @@ double calcEn(std::vector<short>& original, std::vector<short>& noise);
  * @param En energy noise
  * @return Signal-To-Noise ratio
  */
-double calcSNR(long Es, long En);
+double calcSNR(double Es, double En);
 
 /*!
  * Applies a sliding window to a wav SndFile to
